Pre dev (#55)
*refactor: change return message when error on giving score.

* refactor: when uploading audio, check if there is an audio before.

* refactor: add column teacher_name in class table and change getclassdetail function to teachergetclassdetail.

* refactor: add student list in class detail and add get submission detail.

* add: add signedurl when get submission detail.

+++ src/controllers/submission-controller.js
@@ -8,15 +8,38 @@ class SubmissionController {
     const submissionId = req.params.id;
 
     try {
-      const submissionInstance = await submissionModel.get(submissionId);
-      // Delete current audio file
-      await firebaseStorage.deleteFile(
-        submissionInstance.dataValues.audio_file
-      );
+      const submissionInstance = await submissionModel.getById(submissionId);
+      if (submissionInstance.dataValues.audio_file)
+        // Delete current audio file
+        await firebaseStorage.deleteFile(
+          submissionInstance.dataValues.audio_file
+        );
+
       const audioFileName = firebaseStorage.uploadFile(audioFile);
       await submissionModel.uploadAudio(submissionId, audioFileName);
       const response = Response.postResponse(
         "The system successfully uploaded the audio to the submission."
+      );
+      return res.status(200).json(response);
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  static async getSubmissionDetail(req, res, next) {
+    const params = req.query;
+    try {
+      const submissionInstance = await submissionModel.getByTaskAndStudent(
+        params.task_id,
+        params.student_id
+      );
+      const url = await firebaseStorage.getSignedUrl(
+        submissionInstance.dataValues.audio_file
+      );
+      submissionInstance.dataValues.audio_file = url;
+      const response = Response.getResponse(
+        "The system successfully uploaded the audio to the submission.",
+        submissionInstance
       );
       return res.status(200).json(response);
     } catch (error) {
